comment added for intersection module, rename intersection component to intersectionAnimate 2
