+++ pages/Sections/Gift.js
@@ -1,4 +1,4 @@
-import IntersectionAnimate from "../../components/Intersection";
+import IntersectionAnimate from "../../components/IntersectionAnimate";
 
 export default function Gift() {
